Add Door Lock management to all clusters app.
