+++ examples/all-clusters-app/linux/BUILD.gn
@@ -23,6 +23,9 @@ source_set("chip-all-clusters-common") {
     "${chip_root}/examples/all-clusters-app/all-clusters-common/src/binding-handler.cpp",
     "${chip_root}/examples/all-clusters-app/all-clusters-common/src/bridged-actions-stub.cpp",
     "${chip_root}/examples/all-clusters-app/all-clusters-common/src/static-supported-modes-manager.cpp",
+    "${chip_root}/examples/door-lock-app/linux/src/LockEndpoint.cpp",
+    "${chip_root}/examples/door-lock-app/linux/src/LockManager.cpp",
+    "${chip_root}/examples/door-lock-app/linux/src/ZCLDoorLockCallbacks.cpp",
     "include/tv-callbacks.cpp",
     "include/tv-callbacks.h",
     "main-common.cpp",
@@ -34,8 +37,10 @@ source_set("chip-all-clusters-common") {
     "${chip_root}/src/lib",
   ]
 
-  include_dirs =
-      [ "${chip_root}/examples/all-clusters-app/all-clusters-common/include" ]
+  include_dirs = [
+    "${chip_root}/examples/all-clusters-app/all-clusters-common/include",
+    "${chip_root}/examples/door-lock-app/linux/include",
+  ]
 
   cflags = [ "-Wconversion" ]
